requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
PyPDF2
chromadb
pdf2image
faiss-cpu
langchain_google_genai
Pillow
groq 
replicate 
requests